requirements.txt
fastapi
uvicorn[standard]
asyncpg
jinja2
python-multipart
python-dotenv